Renamed installer.py to old_installer.py. No behavior change.
